backend/requirements.txt
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
cmake==4.3.0
colorama==0.4.6
dlib @ file:///E:/4th%20sem/smart_security/dlib-19.24.1-cp311-cp311-win_amd64.whl#sha256=6f1a5ee167975d7952b28e0ce4495f1d9a77644761cf5720fb66d7c6188ae496
face-recognition==1.3.0
face_recognition_models==0.3.0
Flask==3.1.3
flask-cors==6.0.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==1.26.4
opencv-python==4.13.0.92
pillow==12.1.1
requests==2.32.5
urllib3==2.6.3
Werkzeug==3.1.7
